semana3/dia4/requirements.txt
Flask
Flask-SQLAlchemy
mysqlclient
python-dotenv
flask-marshmallow
marshmallow-sqlalchemy